Refactor data structures and update type hints for better clarity and flexibility in strategy calculations

+++ src/app/services/mean_reversion_strategies/optimizer.py
@@ -7,7 +7,7 @@ from app.schemas.internal.best_parameter_combination_dict import (
     ParameterCombinationDict,
     BestResultDict
 )
-from app.services.mean_reversion_strategies.compare_to_buy_and_hold import calculate_comparison_curves
+from app.services.mean_reversion_strategies.strategy_calculations import calculate_comparison_curves
 from app.services.mean_reversion_strategies.mean_reversion_defaults import (
     DEFAULT_INITIAL_CAPITAL,
     DEFAULT_START,
@@ -89,13 +89,13 @@ def optimize_grid_search(
     equity_data = calculate_comparison_curves(best_trades, bot.get_cached_ticker_data(), initial_capital)
 
     return BestParameterCombinationDict(
-        best_drop_threshold=best_params['drop_threshold'],
-        best_hold_days=best_params['hold_days'],
-        best_take_profit_pct=best_params['take_profit_pct'],
-        total_profit=round(best_result['profit'], 2),
-        roi_pct=round(best_roi, 2),
-        win_rate=round(best_result['win_rate'], 2),
-        total_number_of_trades=best_result['total_number_of_trades'],
+        best_drop_threshold=float(best_params['drop_threshold']),
+        best_hold_days=int(best_params['hold_days']),
+        best_take_profit_pct=float(best_params['take_profit_pct']),
+        total_profit=float(round(best_result['profit'], 2)),
+        roi_pct=float(round(best_roi, 2)),
+        win_rate=float(round(best_result['win_rate'], 2)),
+        total_number_of_trades=int(best_result['total_number_of_trades']),
         equity_curve_data=equity_data,
         trades=best_trades
     )
